requirements.txt
telethon==1.32.1
streamlit==1.31.1
python-dotenv==1.0.1 